requirements.txt
numpy
opencv-python
pupil-apriltags
pyntcore